+++ relevanceai/dataset_api/dataset_operations.py
@@ -1751,7 +1751,9 @@ class Operations(Write):
                 dataset_id=self.dataset_id,
                 vector_fields=vector_fields,
             )
-            clusterer.fit_predict_update(dataset=self, vector_fields=vector_fields)
+            clusterer.fit_predict_update(
+                dataset=self, vector_fields=vector_fields, include_grade=True
+            )
 
         elif algorithm.lower() == "hdbscan":
             raise ValueError(